Remove Version and Reasons from submissions UI and database
- Remove Version and Reasons display from submission detail view
- Keep only Description of Change field with edit capability
- Rename ChangeVersion.vue → ChangeDescription.vue
- Clear description field when republishing/unpublishing submissions
- Remove version column from submissions table (use submission_data JSON)
- Update newversion() method to read from submission_data
- Update tests to use submission_data instead of version column
